third_party/blink/web_tests/external/wpt/css/css-sizing/contain-intrinsic-size/auto-006-expected.txt

This is a testharness.js-based test.
PASS Basic usage
PASS Last remembered size can be 0
PASS Last remembered size can be set to 0 after losing display:none
PASS Last remembered size is logical
PASS Last remembered size survives box destruction
PASS Last remembered size survives display type changes
FAIL Losing cis:auto removes last remembered size assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
FAIL Losing cis:auto removes last remembered size even if size doesn't change assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
FAIL Losing cis:auto removes last remembered size immediately assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
PASS Losing cis:auto during display:none doesn't remove last remembered size
FAIL Lack of cis:auto during box creation removes last remembered size assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
FAIL Last remembered size can be removed synchronously assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
PASS Disconnected element can briefly keep last remembered size
FAIL Disconnected element ends up losing last remembered size assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 100
FAIL Disconnected element ends up losing last remembered size even if size was 0x0 assert_equals: Size containment with no last remembered size - clientWidth expected 1 but got 0
FAIL Last remembered size survives becoming inline assert_equals: Still using previous last remembered size - clientWidth expected 100 but got 0
PASS Last remembered size can be set to 0x0 after losing display:inline
Harness: the test ran to completion.
